chore: fetch users on click
