+++ apps/web/src/components/Home/Timeline/index.tsx
@@ -8,7 +8,7 @@ import InfiniteLoader from '@components/UI/InfiniteLoader';
 import type { LensterPublication } from '@generated/types';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { SCROLL_THRESHOLD } from 'data/constants';
-import type { FeedItem, Profile } from 'lens';
+import type { FeedItem } from 'lens';
 import { FeedEventItemType, useTimelineQuery } from 'lens';
 import type { FC } from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
@@ -21,9 +21,6 @@ const Timeline: FC = () => {
   const txnQueue = useTransactionPersistStore((state) => state.txnQueue);
   const feedEventFilters = useTimelinePersistStore((state) => state.feedEventFilters);
   const seeThroughProfile = useTimelineStore((state) => state.seeThroughProfile);
-  const setRecommendedProfilesToSeeThrough = useTimelineStore(
-    (state) => state.setRecommendedProfilesToSeeThrough
-  );
 
   const getFeedEventItems = () => {
     const filters: FeedEventItemType[] = [];
@@ -47,29 +44,8 @@ const Timeline: FC = () => {
   const request = { profileId, limit: 50, feedEventItemTypes: getFeedEventItems() };
   const reactionRequest = currentProfile ? { profileId } : null;
 
-  const setRecommendedProfiles = (feedItems: FeedItem[]) => {
-    let uniqueProfileIds: string[] = [];
-    let profiles: Profile[] = [];
-    for (const feedItem of feedItems) {
-      const profileId = feedItem.root?.profile?.id;
-      if (
-        !uniqueProfileIds.includes(profileId) &&
-        profileId !== seeThroughProfile?.id &&
-        profileId !== currentProfile?.id
-      ) {
-        profiles.push(feedItem.root?.profile as Profile);
-        uniqueProfileIds.push(profileId);
-      }
-    }
-    setRecommendedProfilesToSeeThrough(profiles?.slice(0, 5));
-  };
-
   const { data, loading, error, fetchMore } = useTimelineQuery({
-    variables: { request, reactionRequest, profileId },
-    onCompleted: (data) => {
-      const feedItems = data?.feed?.items as FeedItem[];
-      setRecommendedProfiles(feedItems);
-    }
+    variables: { request, reactionRequest, profileId }
   });
 
   const publications = data?.feed?.items;
